backend/pyproject.toml
[project]
name = "receptra"
version = "0.1.0"
description = "Receptra backend — Hebrew-first local voice co-pilot FastAPI service."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [{ name = "Receptra Contributors" }]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.20",
    # Phase 2 (Hebrew Streaming STT) additions:
    "faster-whisper>=1.2.1,<2",
    "silero-vad>=6.2.1,<7",
    "numpy>=2.0,<3",
    "loguru>=0.7.3,<1",
    # Phase 3 (Hebrew Suggestion LLM) additions:
    "ollama>=0.6.1,<1",
    # Phase 4 (Hebrew RAG Knowledge Base) additions:
    # Thin client only — full `chromadb` package pulls onnxruntime/pulsar-client/
    # tokenizers (RESEARCH §Cluster 1 anti-pattern). Server runs in its own
    # container per docker-compose.yml. Both publish identical 1.5.8.
    "chromadb-client>=1.5.8,<2",
]

[dependency-groups]
dev = [
    "ruff>=0.7",
    "mypy>=1.13",
    "pip-licenses>=5.0",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    # Phase 2 dev/test additions:
    "jiwer>=4.0,<5",
    "soundfile>=0.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/receptra"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]
# Plan 03-02: Hebrew system prompt + few-shot strings are LOCKED RESEARCH §5.2/§5.4
# content. E501 (line length) cannot be wrapped without altering byte content;
# RUF001 (ambiguous Hebrew vs Latin glyphs) is the whole point — the file is Hebrew.
"src/receptra/llm/prompts.py" = ["E501", "RUF001"]
# Hebrew test strings are intentional; RUF001 ambiguous-glyph warnings are noise.
"tests/llm/test_prompts.py" = ["S101", "RUF001"]
# Plan 04-02: Hebrew chunker test fixtures use real Hebrew literals (gershayim
# ״ U+05F4, geresh ׳ U+05F3); RUF001 flags ambiguous-glyph noise on a Hebrew
# corpus. Same precedent as tests/llm/test_prompts.py.
"tests/rag/test_chunker.py" = ["S101", "RUF001", "RUF002", "RUF003"]
# Plan 04-02: chunker.py module docstring documents the gershayim / geresh
# regression contract using the actual Unicode codepoints — the same
# RUF002 (docstring ambiguous-glyph) noise applies as on the test file.
"src/receptra/rag/chunker.py" = ["RUF001", "RUF002", "RUF003"]
# Plan 04-01: ``IngestRejected`` is a contract-locked exception name aligned with
# RESEARCH §REST API ``KbErrorResponse.code`` semantics ("ingest was rejected") —
# the ``Error`` suffix N818 wants would invert past-tense semantics. ``RagInitError``
# already follows the convention; only ``IngestRejected`` triggers N818.
"src/receptra/rag/errors.py" = ["N818"]
# Plan 04-04: Hebrew test strings are intentional; RUF001 ambiguous-glyph warnings are noise.
"tests/rag/test_04_ingest.py" = ["S101", "RUF001"]
"tests/rag/test_04_retriever.py" = ["S101", "RUF001"]

[tool.mypy]
strict = true
python_version = "3.12"
files = ["src", "tests"]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_decorators = false
# Test files intentionally use type: ignore for deliberate type violations
# (e.g. passing str where Literal[...] expected to test runtime validation).
# Pydantic's mypy plugin accepts these without error, so warn_unused_ignores
# would fire spuriously. Disable for test scope only.
warn_unused_ignores = false

# Upstream untyped libraries (no py.typed marker). faster_whisper and
# silero_vad ship without type stubs; their public surface we use is narrow.
[[tool.mypy.overrides]]
module = ["faster_whisper", "faster_whisper.*", "silero_vad", "silero_vad.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
addopts = "-ra --strict-markers"
